+++ CHANGES.md
@@ -17,6 +17,7 @@
 - Renamed short version of ``--lineseparator`` to ``-ls`` to avoid collision with ``--list\-l``
 - Description for disabled transformers can be now displayed & disabled transformers are in ``--list`` output [#114](https://github.com/MarketSquare/robotframework-tidy/issues/114)
 - Robotidy should now correctly load configuration files from path when using ``--config`` [#138](https://github.com/MarketSquare/robotframework-tidy/issues/138)
+- ReplaceRunKeywordIf will now set variable values to `None` if there is no ELSE branch [#140](https://github.com/MarketSquare/robotframework-tidy/issues/140)
 - Transformers should always use the same order. If you need to use custom order, provide --force-order flag [#142](https://github.com/MarketSquare/robotframework-tidy/issues/142)
 
 ### Other